requirements.txt
flask==3.0.1
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
Jinja2==3.1.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-dotenv==1.0.1
PyYAML==5.3.1
redis==5.0.1
requests==2.22.0
simplejson==3.16.0
SQLAlchemy==2.0.25
ssh-import-id==5.10
Twisted==18.9.0
typing-extensions==4.9.0
urllib3==1.25.8
wadllib==1.3.3
werkzeug==3.0.1